+++ src/ui/dock_box.rs
@@ -101,6 +101,7 @@ pub fn build(
     alignment_box: &gtk4::Box,
     ctx: &DockContext,
     win: &gtk4::ApplicationWindow,
+    monitor_name: &str,
 ) -> gtk4::Box {
     let config = &ctx.config;
     let inner_orientation = if config.is_vertical() {
@@ -163,6 +164,12 @@ pub fn build(
         &active_class,
         &ignored_classes,
     );
+
+    if ctx.config.ws {
+        let workspaces_row = crate::ui::workspaces::build_workspace_row(ctx, monitor_name);
+        main_box.append(&workspaces_row);
+    }
+
     build_running_items(
         &main_box,
         ctx,